refactor: Fold smoke.py into diagnostics.py and state.py into types.py
Consolidate Daytona provider: merge smoke validation logic into
diagnostics.py and chat/session normalization helpers into types.py.
Update all internal imports. Net: -2 files.
