+++ src/fleet_rlm/integrations/providers/daytona/agent.py
@@ -10,7 +10,7 @@ from fleet_rlm.runtime.models import StreamEvent
 
 from .interpreter import DaytonaInterpreter
 from .runtime import DaytonaSandboxRuntime
-from .state import (
+from .types import (
     dedupe_paths,
     normalize_history_turn,
 )
